refactor: rename `query-provider.tsx` into `QueryProvider.tsx`
