+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { cn } from '@/lib/utils'
 import { Inter } from 'next/font/google'
 import { StrictMode } from 'react'
 import './globals.css'
-import QueryProvider from '@/components/query-provider'
+import QueryProvider from '@/components/QueryProvider'
 import { Toaster } from '@/components/ui/sonner'
 import type { Metadata } from 'next'
 import { ThemeProvider } from 'next-themes'
